+++ cli/src/args/validate.rs
@@ -16,8 +16,15 @@ pub struct Args {
 impl Run for Args {
     async fn run(self, input: Input, sender: Sender<Value>) -> Result<Option<Value>> {
         let value = input.read(self.infile)?;
-        let result = value.validate();
-        if let Err(stac_validate::Error::Validation(ref errors)) = result {
+        let result = tokio::task::spawn_blocking(move || {
+            if let Err(err) = value.validate() {
+                Err((err, value))
+            } else {
+                Ok(())
+            }
+        })
+        .await?;
+        if let Err((stac_validate::Error::Validation(ref errors), ref value)) = result {
             let message_base = match value {
                 stac::Value::Item(item) => format!("[item={}] ", item.id),
                 stac::Value::Catalog(catalog) => format!("[catalog={}] ", catalog.id),
@@ -32,7 +39,7 @@ impl Run for Args {
                 sender.send(message.into()).await?;
             }
         }
-        result.and(Ok(None)).map_err(Error::from)
+        result.and(Ok(None)).map_err(|(err, _)| Error::from(err))
     }
 
     fn take_outfile(&mut self) -> Option<String> {