+++ cli/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- **tokio** panic when validating ([#358](https://github.com/stac-utils/stac-rs/pull/358))
+
 ## [0.3.0] - 2024-09-05
 
 ### Added